requirements.txt
black==25.1.0
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
flake8==7.2.0
ipython==9.2.0
isort==6.0.1
pillow==11.2.1
psycopg2==2.9.10
python-dotenv==1.1.0
coverage==7.8.2
drf-yasg==1.21.10
drf-spectacular==0.28.0
forex-python==1.9.2
stripe==12.2.0
celery==5.5.3
eventlet==0.40.0
redis==6.2.0
django-celery-beat==2.8.1

